backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic>=2.7.1
httpx==0.27.0
genlayer-py
supabase>=2.0.0
PyPDF2>=3.0.0
